+++ app/services/api/v1/auth_services/get_document/contract.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module Api
+  module V1
+    module AuthServices
+      module GetDocument
+        class Contract < ApplicationContract
+          params do
+            required(:document).filled(:string)
+          end
+
+          rule(:document) do
+            key.failure(I18n.t("api.auth.document")) unless CPF.valid?(values[:document]) || CNPJ.valid?(values[:document])
+          end
+        end
+      end
+    end
+  end
+end