+++ app/services/api/v1/auth_services/get_document/use_case.rb
@@ -1,0 +1,47 @@
+# frozen_string_literal: true
+
+module Api
+  module V1
+    module AuthServices
+      module GetDocument
+        class UseCase < MainService
+          step :validate_params
+          step :find_document
+          step :output
+
+          def validate_params(params)
+            validation = Contract.call(params.permit!.to_h)
+            return Success(params) if validation.success?
+
+            Failure(format_errors(validation.errors.to_h))
+          end
+
+          def find_document(params)
+            document = CNPJ.new(params[:document]).stripped
+            response = fetch_company_data(document)
+
+            return Failure(I18n.t("api.auth.company_not_found")) if response.blank?
+
+            status = response.dig("estabelecimento", "situacao_cadastral")
+            return Failure(I18n.t("api.auth.company_status", status: status)) unless status == "Ativa"
+
+            Success(response)
+          end
+
+          def output(result)
+            response = Presenter.call(result)
+            return Success([ I18n.t("api.auth.company_found"), response ]) if response
+
+            Failure(I18n.t("api.auth.company_not_found"))
+          end
+
+          private
+
+          def fetch_company_data(document)
+            Util.company_data(document)
+          end
+        end
+      end
+    end
+  end
+end