+++ config/application.rb
@@ -34,7 +34,7 @@ module FinanControlApi
     #
     # These settings can be overridden in specific environments using the files
     # in config/environments, which are processed later.
-    #
+    config.i18n.default_locale = :'pt-BR'
     # config.time_zone = "Central Time (US & Canada)"
     # config.eager_load_paths << Rails.root.join("extras")
     config.generators do |generate|